requirements.txt
# Dépendances Python pour le simulateur IoT
paho-mqtt>=1.6.1
flask>=2.3.0
flask-socketio>=5.3.0
flask-cors>=4.0.0
python-socketio>=5.10.0
python-engineio>=4.8.0
simple-websocket>=1.0.0
qrcode>=7.4.2
pillow>=10.0.0
